feat(BE): 면접 종료 시 면접 녹음 및 타임스탬프 저장

See merge request s08-webmobile1-sub2/S08P12D107!149

+++ backend/src/main/java/com/alppano/speakon/domain/recording_timestamp/dto/TimestampWithQuestion.java
@@ -9,6 +9,6 @@ import lombok.*;
 @NoArgsConstructor
 @AllArgsConstructor
 public class TimestampWithQuestion {
-    private int secondTime;
+    private Long secondTime;
     private QuestionSimpleInfo question;
 }
